master - implemented basic structure for masters, customers and their orders

+++ src/main/resources/schema.sql
@@ -30,7 +30,7 @@ CREATE TABLE mvp.masters(
   photo BYTEA,
   active BOOLEAN,
   busy BOOLEAN,
-  rank FLOAT,
+  rank DECIMAL,
   PRIMARY KEY (login, password)
 );
